Add first entry: backstory, issue choice, ramfs
This constitutes the first (incomplete) entry for this project. I still need to add information related to TAP devices, Ethernet vs IP, and iptables.

Also, in the future, my entries will probably be smaller. I want to strike a balance between being informative and being concise, but I'll have to play around with things to get it just right.